Only run CI job in context of upstream repo
When a merge request is opened, the CI pipeline execution context varies

 - If the contributor is a project member the CI pipeline
   runs in context of the upstream / target repo.

 - If the contributor is NOT a project member the CI pipeline
   runs in context of the contributor's fork repo.

The latter is an undesirable situation, because contributing to the
upstream project will burn the contributor's personal CI credits
for shared runners.

By setting a condition

  $CI_PROJECT_PATH == $CI_MERGE_REQUEST_PROJECT_PATH

we ensure the pipeline jobs are never created in context of a
fork repo.

The intended effect is:

  - A project member opening a MR will see a pipeline
    run immediately in upstream context.

  - A non-project member opening a MR will not get any
    pipeline created. A project member can use the MR
    Web UI to manually stasrt a pipeline in upstream
    repo context after verifying the code contribution.

This should work regardless of whether the repo is configured to
run pipelines using 'merged results' or 'detached'.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>